requirements.txt
coloredlogs==15.0.1
ddddocr==1.5.6
flatbuffers==25.12.19
humanfriendly==10.0
mpmath==1.3.0
numpy>=1.26,<2
onnxruntime==1.23.2
opencv-python==4.11.0.86
packaging==26.0
pillow==12.1.1
protobuf==7.34.0
pyreadline3==3.5.4
sympy==1.14.0
python-dotenv==1.0.0
